Reduce connections for very large files
In cases where `$num_chunks` exceeds `$NUM_WORKERS` previously the
excessive downloads were put on hold until download processes fell below
`$NUM_WORKERS`.

In that case this commit reduces `$num_chunks` to be equal to
`$NUM_WORKERS` so that the total number of connections is reduced while
maintaining the desired parallelism.

This also balances the division remainder among all chunks so that the
last file is never more than `$num_workers` larger than the others,
which is a further optimization.