Modyfied install steps with notes that worked for me.
Thanks for the porject, everyone. I just went through the install process
and thought maybe the mac directions could use a tiny bit of clarification.
I'm by no means familiar enough with the project to know if this is the best way
to convey my experience but I figured I'd give it a shot and maybe it will help
someone in the future.